Create request_handler app
In this commit a request_handler app is created to begin handling requests from students.
It includes a Request Model, a RequestForm ModelForm, and the views to allow them to be used.
For now, it is possible to create a request from the Dashboard. If successful, the user is redirected a success page.

Although the request is internally processed, it is essentially "lost" to the user, as there is still some functionality to be implemented.