Add "Availability" and "Mode Preference" fields to request form.
Two new fields are added to the form. They take the shape of checkboxes in which the user can select 1)The days of the week in which they are available; 2)The mode of delivery they prefer(in person, online, or no preference).

Also begin adding test file to the app.