Enhance test suite to achieve 99% code coverage.
More tests have been added to the test suite. These include more tests on test_model.py and a new set of tests in test_views.py. The latter aims to produce tests for the views defined in this app, while the former adds additional tests to the Request model.